(feat) Updated buffer gas amount for fee in all examples
