+++ examples/chain_client/31_MsgCreateBinaryOptionsLimitOrder.py
@@ -2,9 +2,10 @@ import asyncio
 import uuid
 
 from pyinjective.async_client import AsyncClient
-from pyinjective.constant import Denom
+from pyinjective.constant import GAS_FEE_BUFFER_AMOUNT, GAS_PRICE
 from pyinjective.core.network import Network
 from pyinjective.transaction import Transaction
+from pyinjective.utils.denom import Denom
 from pyinjective.wallet import PrivateKey
 
 
@@ -68,8 +69,8 @@ async def main() -> None:
     print(sim_res_msg)
 
     # build tx
-    gas_price = 500000000
-    gas_limit = sim_res.gas_info.gas_used + 20000  # add 20k for gas, fee computation
+    gas_price = GAS_PRICE
+    gas_limit = sim_res.gas_info.gas_used + GAS_FEE_BUFFER_AMOUNT  # add buffer for gas fee computation
     gas_fee = "{:.18f}".format((gas_price * gas_limit) / pow(10, 18)).rstrip("0")
     fee = [
         composer.Coin(
