+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "injective-py"
-version = "1.0"
+version = "1.0-rc1"
 description = "Injective Python SDK, with Exchange API Client"
 authors = ["Injective Labs <contact@injectivelabs.org>"]
 license = "Apache-2.0"
